+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Adding a script to setup the Azure Batch Pool to link the container.
 * Adding new action to post a comment on PRs with a link to the rendered pkgdown site.
+* Add inner pipeline responsible for running the model fitting process
 * Re-organizing GitHub workflows.
 * Checks if batch pool exists. Pools are named after branches. Also allows for deletion via commit message.
 * Merges workflows 1 and 2 into a single workflow.